Rename ChipInstance::synth_write_reg(), disable writing to ARAM
Turns out S-DSP registers aren't directly memory-mapped,
but have separate memory addresses for selecting a register
and sending a value.

+++ src/audio/synth/music_driver_common.h
@@ -23,7 +23,7 @@ struct TimeRef {
 template<int begin, int end, typename T>
 using Range = T;
 
-using Address = uint16_t;
+using Address = uint8_t;
 using Byte = uint8_t;
 
 struct RegisterWrite {
